IGNITE-14477 Ducktape test of rebalance for persistent mode
